split temperature protection into heating and cooling (#30662)
* split temperature protection into heating and cooling

* fugg

* weh
