+++ Resources/Prototypes/Entities/Clothing/OuterClothing/wintercoats.yml
@@ -46,7 +46,8 @@
   - type: Clothing
     sprite: Clothing/OuterClothing/WinterCoats/coat.rsi
   - type: TemperatureProtection
-    coefficient: 0.1
+    heatingCoefficient: 1.1
+    coolingCoefficient: 0.1
   - type: Item
     size: Normal
   - type: Armor
