+++ Resources/Prototypes/Entities/Clothing/Shoes/base_clothingshoes.yml
@@ -91,4 +91,5 @@
   description: Fluffy boots to help survive even the coldest of winters.
   components:
   - type: TemperatureProtection
-    coefficient: 0.8
+    heatingCoefficient: 1.025
+    coolingCoefficient: 0.5
